+++ src/elements/image.rs
@@ -1,0 +1,1 @@
+use crate::bindings::*;
