+++ src/elements/text.rs
@@ -1,0 +1,9 @@
+use crate::bindings::*;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+#[repr(u32)]
+pub enum TextElementConfigWrapMode {
+    Words = Clay_TextElementConfigWrapMode_CLAY_TEXT_WRAP_WORDS,
+    Newline = Clay_TextElementConfigWrapMode_CLAY_TEXT_WRAP_NEWLINES,
+    None = Clay_TextElementConfigWrapMode_CLAY_TEXT_WRAP_NONE,
+}
